- Renamed a lot of the core network functions to use the c++ std naming format. - Most of the libs and servers have been updated to use the new networking system.

+++ src/core/include/network_thread_pool.h
@@ -41,7 +41,7 @@ class NetworkThreadPool {
   }
 
   asio::io_context* Get_IO_Service() { return &io_service_; }
-  uint16_t GetThreadCount() const { return threads_active_.count(); }
+  uint16_t GetThreadCount() const { return static_cast<uint16_t>(threads_active_.count()); }
 
  private:
   NetworkThreadPool(uint16_t maxthreads) : io_work_(new asio_worker::element_type(io_service_)) {
